Whatever

+++ Include/Rr/Rr_Image.h
@@ -131,7 +131,7 @@ extern Rr_Image2D *Rr_CreateImage2D(
 
 extern Rr_Image2DArray *Rr_CreateImage2DArray(
     Rr_IntVec2 Extent,
-    size_t ArrayCount,
+    uint32_t ArrayCount,
     Rr_TextureFormat Format,
     Rr_ImageFlags Flags);
